Move bundle hashing to web plugin
