+++ build.sbt
@@ -32,31 +32,31 @@ lazy val commonSettings = Seq(
 lazy val `sbt-scalajs-esbuild` =
   project
     .in(file("sbt-scalajs-esbuild"))
-    .enablePlugins(SbtPlugin, ShadingPlugin)
+    .enablePlugins(SbtPlugin)
     .settings(commonSettings)
     .settings(
-      addSbtPlugin("org.scala-js" % "sbt-scalajs" % "1.10.1"),
-      libraryDependencies += "org.typelevel" %% "jawn-ast" % "1.4.0",
-      shadedModules ++= Set(
-        "org.typelevel" %% "jawn-ast",
-        "org.typelevel" %% "jawn-parser",
-        "org.typelevel" %% "jawn-util"
-      ),
-      shadingRules ++= Seq(
-        ShadingRule
-          .moveUnder(
-            "org.typelevel.jawn",
-            "scalajsesbuild.shaded.org.typelevel.jawn"
-          )
-      ),
-      validNamespaces ++= Set("scalajsesbuild", "sbt")
+      addSbtPlugin("org.scala-js" % "sbt-scalajs" % "1.10.1")
     )
 
 lazy val `sbt-scalajs-esbuild-web` = project
   .in(file("sbt-scalajs-esbuild-web"))
-  .enablePlugins(SbtPlugin)
+  .enablePlugins(SbtPlugin, ShadingPlugin)
   .settings(
     commonSettings,
+    libraryDependencies += "org.typelevel" %% "jawn-ast" % "1.4.0",
+    shadedModules ++= Set(
+      "org.typelevel" %% "jawn-ast",
+      "org.typelevel" %% "jawn-parser",
+      "org.typelevel" %% "jawn-util"
+    ),
+    shadingRules ++= Seq(
+      ShadingRule
+        .moveUnder(
+          "org.typelevel.jawn",
+          "scalajsesbuild.shaded.org.typelevel.jawn"
+        )
+    ),
+    validNamespaces ++= Set("scalajsesbuild", "sbt"),
     scriptedDependencies := {
       val () = scriptedDependencies.value
       val () = (`sbt-scalajs-esbuild` / publishLocal).value
